src/main/java/com/tbawor/jrogal/game/hero/Statistics.java
/**
 * MIT License
 *
 * Copyright (c) 2017 Tomasz Bawor
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */
package com.tbawor.jrogal.game.hero;

/**
 * Statistics of hero.
 * @author Tomasz Bawor (bawortomasz@gmail.com)
 * @version $Id$
 * @since 0.0.1
 */
final class Statistics {

    /**
     * Hero Attack value.
     */
    private final int attack;

    /**
     * Hero Defence value.
     */
    private final int defence;

    /**
     * Constructor for creating statistics.
     * @param attack Attack int value
     * @param defence Defence int value
     */
    Statistics(final int attack, final int defence) {
        this.attack = attack;
        this.defence = defence;
    }

    /**
     * Method returning attack value.
     * @return Attack value
     */
    public int getAttack() {
        return this.attack;
    }

    /**
     * Mehtod returning defence value.
     * @return Defence value
     */
    public int getDefence() {
        return this.defence;
    }
}
